// types/IResponse.ts
export interface IResponse<T> {
  success: boolean;
  message: string;
  data: T | null;
  error: string | null;

}

// Define the response type for the backend
export interface ApiResponse<T> {
  success: boolean;
  message: string;
  data?: T | null; // ใช้ประเภทเจนเนอริกที่นี่
  error?: string;
  token?: string; // Add token to the ApiResponse type
  

}

export interface ApiResponseS<T> {
  success: boolean;
  totalCount: number; // Include totalCount here
  message: string;
  data?: T | null; // ใช้ประเภทเจนเนอริกที่นี่
  error?: string;
  token?: string; // Add token to the ApiResponse type

}

export interface ApiResponseI<T> {
  success: boolean; // Indicates if the API request was successful
  message: string; // Contains a message, usually for success or error details
  data?: T | null; // Holds the data returned from the API (generic type), which can also be null
  error?: string; // Error message in case the request fails
}